requirements.txt
asgiref==3.2.10
Django==3.1.1
djangorestframework==3.11.1
Dumper==1.2.0
Faker==4.1.3
gunicorn==20.0.4
pkg-resources==0.0.0
python-dateutil==2.8.1
pytz==2020.1
randomtimestamp==1.0.0
six==1.15.0
sqlparse==0.3.1
text-unidecode==1.3
